+++ supvisors/sparser.py
@@ -17,7 +17,6 @@
 # limitations under the License.
 # ======================================================================
 
-from collections import OrderedDict
 from distutils.util import strtobool
 from os import path
 from sys import stderr
@@ -53,22 +52,25 @@ class Parser(object):
         self.logger = supvisors.logger
         self.tree = self.parse(supvisors.options.rules_file)
         self.root = self.tree.getroot()
+        # get aliases - check nodes and back to string as it is easier to process
+        self.aliases = {element.get('name'): list_of_strings(element.text)
+                        for element in self.root.findall("./alias[@name]")
+                        if element.text}
+        self.logger.debug('Parser: found aliases {}'.format(self.aliases))
         # get models
         elements = self.root.findall("./model[@name]")
         self.models = {element.get('name'): element for element in elements}
-        self.logger.debug('Parser.__init__: found models {}'.format(self.models.keys()))
+        self.logger.debug('Parser: found models {}'.format(self.models.keys()))
         # get application patterns
         app_elements = self.root.findall(".//application[@pattern]")
         self.application_patterns = {app_element.get('pattern'): app_element for app_element in app_elements}
         self.logger.debug('Parser: found application patterns {}'.format(self.application_patterns.keys()))
         # get program patterns
-        # found program patterns {None: dict_keys(['', 'check_', 'internal_data_bus'])}
         self.program_patterns = {}
         app_elements = self.root.findall(".//application/pattern[@name]/..")
         for app_element in app_elements:
             prg_elements = app_element.findall("./pattern[@name]")
-            self.program_patterns[app_element] = {prg_element.get('name'): prg_element
-                                                  for prg_element in prg_elements}
+            self.program_patterns[app_element] = {prg_element.get('name'): prg_element for prg_element in prg_elements}
         if self.program_patterns:
             self.logger.warn('Parser: usage of pattern elements is deprecated -'
                              ' please convert {} to program elements with pattern attribute.'
@@ -85,7 +87,7 @@ class Parser(object):
 
         :return: a list of program patterns per application name or patterns
         """
-        return {app_element.get('name') or app_element.get('pattern'): list(prg_patterns.keys())
+        return {Parser.get_element_name(app_element): list(prg_patterns.keys())
                 for app_element, prg_patterns in self.program_patterns.items()}
 
     def load_application_rules(self, application_name: str, rules: ApplicationRules) -> None:
@@ -171,8 +173,8 @@ class Parser(object):
         # find check if a model is referenced in the program rules
         model_elt = self.get_model_element(program_elt)
         if model_elt is not None:
-            self.logger.trace('Parser.load_model_rules: found model {} from program {}'
-                              .format(model_elt.get('name'), program_elt.get('name')))
+            self.logger.trace('Parser.load_model_rules: found model={} from program={}'
+                              .format(model_elt.get('name'), Parser.get_element_name(program_elt)))
             # a model can reference another model
             # WARN: recursive call, counter decreased
             self.load_model_rules(model_elt, rules, loop_check - 1)
@@ -184,6 +186,15 @@ class Parser(object):
         self.load_boolean(program_elt, 'wait_exit', rules)
         self.load_expected_loading(program_elt, rules)
         self.load_enum(program_elt, 'running_failure_strategy', RunningFailureStrategies, rules)
+
+    @staticmethod
+    def get_element_name(elt: Any):
+        """ Return the name or the pattern name of the element.
+
+        :param elt: the application or program element
+        :return: the name or pattern name attribute of the element
+        """
+        return elt.get('name') or elt.get('pattern')
 
     def get_program_element(self, namespec: str) -> Optional[Any]:
         """ Try to find the definition of a program in rules files.
@@ -227,6 +238,39 @@ class Parser(object):
         model = elt.findtext('reference')
         return self.models.get(model, None)
 
+    def check_node_list(self, str_node_list: str):
+        """ Resolve and check the list of nodes provided.
+
+        :param str_node_list: the node names, separated by commas
+        :return: the list of validated nodes and of validated hash_nodes
+        """
+        # resolve aliases
+        # Version 1: simple pass on input elements considering that an alias cannot include another alias
+        # node_names = []
+        # for node_name in list_of_strings(str_node_list):
+        #     if node_name in self.aliases:
+        #         node_names.extend(self.aliases[node_name])
+        #     else:
+        #         node_names.append(node_name)
+        # Version 2: use list slicing to insert aliases
+        # here an alias can be referenced in another alias if declared after in the XML
+        node_names = list_of_strings(str_node_list)
+        for alias_name, alias in self.aliases.items():
+            if alias_name in node_names:
+                pos = node_names.index(alias_name)
+                node_names[pos:pos] = alias
+        # keep reference to hashtag as it will be removed by the filters
+        ref_hashtag = '#' in node_names
+        if '*' in node_names:
+            node_names = ['*']
+        else:
+            # filter the unknown nodes (or remaining aliases)
+            node_names = self.supvisors.address_mapper.filter(node_names)
+        # re-inject the hashtag if needed. position does not matter
+        if ref_hashtag:
+            node_names.append('#')
+        return node_names
+
     def load_application_nodes(self, elt: Any, rules: ApplicationRules) -> None:
         """ Get the nodes where the non-distributed application is authorized to run.
 
@@ -236,12 +280,7 @@ class Parser(object):
         """
         value = elt.findtext('addresses')
         if value:
-            # sort and trim list of values
-            node_names = list(OrderedDict.fromkeys(filter(None, list_of_strings(value))))
-            if '*' in node_names:
-                rules.node_names = ['*']
-            else:
-                rules.node_names = self.supvisors.address_mapper.filter(node_names)
+            rules.node_names = self.check_node_list(value)
 
     def load_program_nodes(self, elt: Any, rules: ProcessRules) -> None:
         """ Get the nodes where the program is authorized to run.
@@ -252,22 +291,19 @@ class Parser(object):
         """
         value = elt.findtext('addresses')
         if value:
-            # sort and trim list of values
-            node_names = list(OrderedDict.fromkeys(filter(None, list_of_strings(value))))
-            if '#' in node_names:
-                # process cannot be started anywhere until hash node is resolved
-                rules.node_names = []
+            rules.node_names = self.check_node_list(value)
+            if '#' in rules.node_names:
                 # if '#' is alone or associated to '*', the logic is applicable to all addresses
-                if len(node_names) == 1 or '*' in node_names:
+                if len(rules.node_names) == 1 or '*' in rules.node_names:
                     rules.hash_node_names = ['*']
                 else:
                     # '#' is applicable to a subset of node names
-                    node_names.remove('#')
-                    rules.hash_node_names = node_names
-            elif '*' in node_names:
+                    rules.node_names.remove('#')
+                    rules.hash_node_names = rules.node_names
+                # process cannot be started anywhere until hash node is resolved
+                rules.node_names = []
+            elif '*' in rules.node_names:
                 rules.node_names = ['*']
-            else:
-                rules.node_names = self.supvisors.address_mapper.filter(node_names)
 
     def load_sequence(self, elt: Any, attr_string: str, rules: AnyRules) -> None:
         """ Return the sequence value found from the XML element.
@@ -285,11 +321,11 @@ class Parser(object):
                 if value >= 0:
                     setattr(rules, attr_string, value)
                 else:
-                    self.logger.warn('Parser.load_sequence: invalid value for {} {}: {} (expected integer >= 0)'
-                                     .format(elt.get('name'), attr_string, value))
+                    self.logger.error('Parser.load_sequence: invalid value for elt={} {}: {} (expected integer >= 0)'
+                                      .format(Parser.get_element_name(elt), attr_string, value))
             except (TypeError, ValueError):
-                self.logger.warn('Parser.load_sequence: not an integer for {} {}: {}'
-                                 .format(elt.get('name'), attr_string, str_value))
+                self.logger.error('Parser.load_sequence: not an integer for elt={} {}: {}'
+                                  .format(Parser.get_element_name(elt), attr_string, str_value))
 
     def load_expected_loading(self, elt: Any, rules: ProcessRules) -> None:
         """ Return the expected_loading value found from the XML element.
@@ -306,11 +342,11 @@ class Parser(object):
                 if 0 <= value <= 100:
                     setattr(rules, 'expected_load', value)
                 else:
-                    self.logger.warn('Parser.load_expected_loading: invalid value for {} expected_loading: {}'
-                                     '(expected integer in [0;100])'.format(elt.get('name'), value))
+                    self.logger.warn('Parser.load_expected_loading: invalid value for elt={} expected_loading: {}'
+                                     '(expected integer in [0;100])'.format(Parser.get_element_name(elt), value))
             except (TypeError, ValueError):
-                self.logger.warn('Parser.load_expected_loading: not an integer for {} expected_loading: {}'
-                                 .format(elt.get('name'), str_value))
+                self.logger.warn('Parser.load_expected_loading: not an integer for elt={} expected_loading: {}'
+                                 .format(Parser.get_element_name(elt), str_value))
 
     def load_boolean(self, elt: Any, attr_string: str, rules: AnyRules) -> None:
         """ Return the boolean value found from XML element.
@@ -326,8 +362,8 @@ class Parser(object):
                 value = bool(strtobool(str_value))
                 setattr(rules, attr_string, value)
             except ValueError:
-                self.logger.warn('Parser.load_boolean: not a boolean-like for {} {}: {}'
-                                 .format(elt.get('name'), attr_string, str_value))
+                self.logger.warn('Parser.load_boolean: not a boolean-like for elt={} {}: {}'
+                                 .format(Parser.get_element_name(elt), attr_string, str_value))
 
     def load_enum(self, elt: Any, attr_string: str, klass: EnumClassType, rules: AnyRules) -> None:
         """ Return the running_failure_strategy value found from XML element.
@@ -344,8 +380,8 @@ class Parser(object):
             try:
                 setattr(rules, attr_string, klass[value])
             except KeyError:
-                self.logger.warn('Pattern.load_enum: invalid value for {} {}: {} (expected in {})'
-                                 .format(elt.get('name'), attr_string, value, klass._member_names_))
+                self.logger.warn('Pattern.load_enum: invalid value for elt={} {}: {} (expected in {})'
+                                 .format(Parser.get_element_name(elt), attr_string, value, klass._member_names_))
 
     def parse(self, filename: str) -> Optional[Any]:
         """ Parse the file depending on the modules installed.